flavie's comments
