+++ docs/releases/changelog-dev.md
@@ -7,6 +7,9 @@
 
 ### Improvements
 
+- Added the `_Sentinels` enum class in `utils/sentinels.py`, currently exposing only `UNSET` (more sentinels can be added later). Sentinels mark uninitialized, unset or otherwise undefined values without relying on `None`, which keeps the logic clearer and makes it possible to distinguish an explicit `None` from an unset default.
+  [#1173](https://github.com/qilimanjaro-tech/qililab/pull/1173)
+
 ### Breaking changes
 
 ### Deprecations / Removals
@@ -15,5 +18,5 @@
 
 ### Bug fixes
 
-- Make it so if you input None in `NonLinearCrosstalkMatrix.set_non_linear_params` it sets the parameters to None, instead of ignoring them.
+- Passing `None` to `NonLinearCrosstalkMatrix.set_non_linear_params` now clears the parameters instead of keeping their previously set values, allowing users to remove non-linear parameters that were set earlier.
   [#1173](https://github.com/qilimanjaro-tech/qililab/pull/1173)
