+++ src/qililab/utils/__init__.py
@@ -20,12 +20,15 @@ from .coordinate_decomposition import coordinate_decompose
 from .dictionaries import merge_dictionaries
 from .factory import Factory
 from .nested_dict_iterator import nested_dict_to_pandas_dataframe
+from .sentinels import Unset, _Sentinel
 from .singleton import Singleton, SingletonABC
 
 __all__ = [
     "Factory",
     "Singleton",
     "SingletonABC",
+    "Unset",
+    "_Sentinel",
     "argsort_buses",
     "coordinate_decompose",
     "dict_factory",
